[FEATURE] added tests for models and bugfixed accordingly

+++ database/migrations/2025_09_25_120632_create_target_demographics_table.php
@@ -18,11 +18,11 @@ return new class extends Migration
                 ->nullable();
             $table->integer('amount')
                 ->default(0);
-            $table->integer('difficulty')
+            $table->float('difficulty')
                 ->default(0);
             $table->string('ethics')
                 ->nullable();
-            $table->integer('relevance')
+            $table->float('relevance')
                 ->default(0);
             $table->timestamps();
         });
